requirements.txt
streamlit
openai
google-generativeai
elasticsearch
pymupdf
python-dotenv
pandas
numpy
aiofiles
requests
pydantic
python-multipart